fixed up download code
Before, the download code had several issues. First, it did not support the "Range" header, so resuming downloads or streaming videos (in Safari, at least) did not work. Now, it uses http.ServeContent(), so video streaming works.

Before, HandleView() and HandleDownload() had a lot of shared, reused, redundant code. Now, it is all stubbed out to one serveFile() function. In addition, the code used to have a bunch of unhelpful comments. In general, I sucked at code when I wrote this. Now the download code is nicer; I cannot speak for the rest of the code.